+++ src/prompter/Prompter.js
@@ -7,16 +7,20 @@ export default class Prompter {
     this.editor = editor;
     this.wordToChange = '';
     this.candidate = '';
+    this.lastWord = '';
 
     this.prompterElement = document.createElement('div');
     // bdi used to get punctuation to the right using direction:rtl
-    this.bdiElement = document.createElement('bdi');
+    this.textElement = document.createElement('div');
+    this.textElement.id = 'prompter-text';
     this.prompterElement.id = 'prompter';
     this.prompterElement.classList.add('prompter');
-    this.prompterElement.appendChild(this.bdiElement);
+    this.prompterElement.setAttribute('touch-action', 'none');
+    this.prompterElement.appendChild(this.textElement);
 
     this.ellipsisElement = document.createElement('div');
     this.ellipsisElement.id = 'ellipsis';
+    this.ellipsisElement.setAttribute('touch-action', 'none');
     this.ellipsisElement.innerHTML = '...';
 
     this.candidatesElement = document.createElement('div');
@@ -35,7 +39,7 @@ export default class Prompter {
 
   addListeners() {
     // We use bdi element and not prompter to only get text
-    this.bdiElement.addEventListener('click', this.showCandidates.bind(this), false);
+    this.textElement.addEventListener('click', this.showCandidates.bind(this), false);
     this.candidatesElement.addEventListener('click', this.clickCandidate.bind(this), false);
     this.ellipsisElement.addEventListener('click', this.showOptions.bind(this), false);
     this.convertElement.addEventListener('click', () => {
@@ -44,70 +48,67 @@ export default class Prompter {
   }
 
   showOptions(evt) {
-    logger.debug(this.ellipsisElement);
-    logger.debug(evt);
-    const id = evt.target.id;
+    if (this.candidatesElement.style.display !== 'none') {
+      this.candidatesElement.style.display = 'none';
+    }
+    if (this.optionsElement.style.display !== 'none') {
+      this.optionsElement.style.display = 'none';
+    } else {
+      this.optionsElement.style.display = 'block';
 
-    this.optionsElement.style.display = 'block';
+      const top = evt.target.getBoundingClientRect().top + 47;
+      const left = evt.target.getBoundingClientRect().left - 30;
+      this.optionsElement.style.top = `${top}px`;
+      this.optionsElement.style.left = `${left}px`;
 
-    const top = evt.target.getBoundingClientRect().top + 32;
-    const left = evt.target.getBoundingClientRect().left - 42;
-    this.optionsElement.style.top = `${top}px`;
-    this.optionsElement.style.left = `${left}px`;
+      this.optionsElement.appendChild(this.convertElement);
 
-    this.optionsElement.appendChild(this.convertElement);
-
-
-    const parent = evt.target.parentNode;
-    parent.insertBefore(this.optionsElement, evt.target);
+      const parent = evt.target.parentNode;
+      parent.insertBefore(this.optionsElement, evt.target);
+    }
   }
 
   showCandidates(evt) {
-    const id = evt.target.id;
-    const words = JSON.parse(this.editor.exports['application/vnd.myscript.jiix']).words;
-    this.wordToChange = words[id];
-    logger.debug(this.wordToChange);
-    this.wordToChange.id = id;
-    this.candidatesElement.innerHTML = '';
-    if (this.wordToChange && this.wordToChange.candidates) {
-      this.candidatesElement.style.display = 'block';
-      this.wordToChange.candidates.forEach((word, index) => {
-        if (this.wordToChange.label === word) {
-          this.candidatesElement.innerHTML += `<span id=${index}><b>${word}</b> &#10004;</span><br>`;
-        } else {
-          this.candidatesElement.innerHTML += `<span id=${index}>${word}</span><br>`;
-        }
-      });
-      // get the parent parent of word to insert just before prompter
-      // 32 to get the boundary of prompter element
-      const top = evt.target.getBoundingClientRect().top + 32;
-      this.candidatesElement.style.top = `${top}px`;
-      this.candidatesElement.style.left = `${evt.target.getBoundingClientRect().left}px`;
+    if (this.optionsElement.style.display !== 'none') {
+      this.optionsElement.style.display = 'none';
+    } else {
+      const id = evt.target.id;
+      const words = JSON.parse(this.editor.exports['application/vnd.myscript.jiix']).words;
+      this.wordToChange = words[id];
+      this.wordToChange.id = id;
+      this.candidatesElement.innerHTML = '';
+      if (this.wordToChange && this.wordToChange.candidates) {
+        this.candidatesElement.style.display = 'block';
+        this.wordToChange.candidates.forEach((word, index) => {
+          if (this.wordToChange.label === word) {
+            this.candidatesElement.innerHTML += `<span id=${index}><b>${word}</b> &#10004;</span><br>`;
+          } else {
+            this.candidatesElement.innerHTML += `<span id=${index}>${word}</span><br>`;
+          }
+        });
+        // get the parent parent of word to insert just before prompter
+        // 32 to get the boundary of prompter element
+        const top = evt.target.getBoundingClientRect().top + 32;
+        this.candidatesElement.style.top = `${top}px`;
+        this.candidatesElement.style.left = `${evt.target.getBoundingClientRect().left}px`;
 
-      const parent = evt.target.parentNode.parentNode.parentNode;
-      parent.insertBefore(this.candidatesElement, evt.target.parentNode.parentNode);
+        const parent = evt.target.parentNode.parentNode.parentNode;
+        parent.insertBefore(this.candidatesElement, evt.target.parentNode.parentNode);
+      }
     }
   }
 
   clickCandidate(evt) {
-    logger.debug('click candidate');
     this.candidate = evt.target.innerText;
-    logger.debug(this.wordToChange);
-    logger.debug(this.candidate);
-    const jiixToImport = JSON.parse(this.editor.exports['application/vnd.myscript.jiix']);
-    jiixToImport.words[this.wordToChange.id].label = this.candidate;
-    const xToImport = jiixToImport.words[0]['bounding-box'].x;
-    const yToImport = jiixToImport.words[0]['bounding-box'].y;
-    logger.debug(jiixToImport);
-    this.editor.importContent({ x: xToImport, y: yToImport }, JSON.stringify(jiixToImport), 'application/vnd.myscript.jiix');
-
+    if (this.candidate !== this.wordToChange.label && this.wordToChange.candidates.includes(this.candidate)) {
+      const jiixToImport = JSON.parse(this.editor.exports['application/vnd.myscript.jiix']);
+      jiixToImport.words[this.wordToChange.id].label = this.candidate;
+      const xToImport = jiixToImport.words[0]['bounding-box'].x;
+      const yToImport = jiixToImport.words[0]['bounding-box'].y;
+      this.editor.importContent({ x: xToImport, y: yToImport }, JSON.stringify(jiixToImport), 'application/vnd.myscript.jiix');
+    }
     this.hideCandidates();
   }
-
-  /*
-    clickOption(evt) {
-      logger.debug('click option');
-    } */
 
   hideCandidates() {
     this.candidatesElement.style.display = 'none';
@@ -118,42 +119,51 @@ export default class Prompter {
   }
 
   populatePrompter(exports) {
-    this.insertPrompter(exports);
-    const exportsJiix = JSON.parse(exports['application/vnd.myscript.jiix']);
-    logger.debug(exportsJiix);
-    const words = exportsJiix.words;
-    this.bdiElement.innerHTML = '<span class="paragraph-icon">&#182;</span>';
+    if (exports) {
+      this.insertPrompter(exports);
+      const exportsJiix = JSON.parse(exports['application/vnd.myscript.jiix']);
+      const words = exportsJiix.words;
 
-    words.forEach((word, index) => {
-      if (word.label === ' ') {
-        this.bdiElement.innerHTML += `<span id=${index}>&nbsp;</span>`;
-      } else {
-        this.bdiElement.innerHTML += `<span id=${index}>${word.label}</span>`;
+      this.textElement.innerHTML = '<span class="paragraph-icon">&#182;</span>';
+
+      words.forEach((word, index) => {
+        if (word.label === ' ') {
+          this.textElement.innerHTML += `<span id=${index}>&nbsp;</span>`;
+        } else {
+          this.textElement.innerHTML += `<span id=${index}>${word.label}</span>`;
+          if (index === words.length - 1) {
+            if (this.lastWord === '') {
+              this.lastWord = word;
+            }
+            // This is used to scroll to last word if last word is modified
+            if (JSON.stringify(this.lastWord) !== JSON.stringify(word)) {
+              document.getElementById(index).scrollIntoView({ behavior: 'smooth' });
+              this.lastWord = word;
+            }
+          }
+        }
+      });
+
+      if (exportsJiix['bounding-box']) {
+        const mm = 3.779528;
+        const top = exportsJiix['bounding-box'].y < 15 ? (exportsJiix['bounding-box'].y * mm) + 2 : (exportsJiix['bounding-box'].y * mm) - 12;
+        const left = (exportsJiix['bounding-box'].x * mm) + this.prompterElement.offsetWidth;
+        this.insertEllipsis(left, top);
       }
-    });
-
-    if (exportsJiix['bounding-box']) {
-      const mm = 3.779528;
-      let top = 0;
-      top = exportsJiix['bounding-box'].y < 15 ? (exportsJiix['bounding-box'].y * mm) + 2 : (exportsJiix['bounding-box'].y * mm) - 12;
-      const left = (exportsJiix['bounding-box'].x * mm) + this.prompterElement.offsetWidth;
-      this.insertEllipsis(left, top);
+      this.hideCandidates();
+      this.hideOptions();
+    } else {
+      this.clearPrompter();
     }
-
-    // Ensure candidates are hidden
-    this.hideCandidates();
-    this.hideOptions();
   }
 
   insertPrompter(exports) {
     const exportsJiix = JSON.parse(exports['application/vnd.myscript.jiix']);
-    const prompterElement = document.querySelector('#prompter');
-    const ellipsisElement = document.querySelector('#ellipsis');
-
-    if (prompterElement && exportsJiix.words.length === 0) {
-      prompterElement.parentNode.removeChild(prompterElement);
-      ellipsisElement.parentNode.removeChild(ellipsisElement);
+    if (this.prompterElement && exportsJiix.words.length === 0) {
+      this.clearPrompter();
     } else if (exportsJiix['bounding-box']) {
+      this.prompterElement.style.display = 'initial';
+      this.ellipsisElement.style.display = 'initial';
       const mm = 3.779528;
       let top = 0;
       // Ensure that the prompter is not placed over the nav bar when text is converted
@@ -173,6 +183,11 @@ export default class Prompter {
         parent.insertBefore(this.prompterElement, this.editor.domElement);
       }
     }
+  }
+
+  clearPrompter() {
+    this.prompterElement.style.display = 'none';
+    this.ellipsisElement.style.display = 'none';
   }
 
   insertEllipsis(left, top) {